services/model-service/requirements.txt
# Model Service Requirements
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.1
pydantic>=2.0.0

# ML dependencies
torch>=2.0.0
pandas>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.21.0 